Create ProjectX.mq5
Hi! This is my first project as I start learning programming through MQL5, since trading is a field I’m most familiar and interested in. The goal of this project is to build a small platform for testing trading patterns based on specific rules. Focus is placed on flexibility and scalability for easier addition of new patterns, filters, and variables. The project is still in development.